requirements.txt
aiogram==3.13.0
python-dotenv==1.0.1
pydantic==2.7.0
aiofiles==23.2.1 